+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/2.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- option to automatically delete downloaded files after a book is marked as finished (off, immediately, or after a configurable delay)
+
 ## [v0.10.2-beta.1] - 2026-09-23
 
 ### Added